+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/package-info.java
@@ -41,7 +41,6 @@
  * <li>{@link is.codion.swing.framework.ui.EntityPanel.Config#CONTROL_TOOLBAR_CONSTRAINTS}
  * <li>{@link is.codion.swing.framework.ui.EntityPanel.Config#WINDOW_TYPE}
  * <li>{@link is.codion.swing.framework.ui.EntityTablePanel.Config#QUERY_HIDDEN_COLUMNS}
- * <li>{@link is.codion.swing.framework.ui.EntityTablePanel.Config#ORDER_QUERY_BY_SORT_ORDER}
  * <li>{@link is.codion.swing.framework.ui.EntityTablePanel.Config#CONDITION_VIEW}
  * <li>{@link is.codion.swing.framework.ui.EntityTablePanel.Config#FILTER_VIEW}
  * <li>{@link is.codion.swing.framework.ui.EntityTablePanel.Config#SUMMARY_PANEL_VISIBLE}