+++ changelog.md
@@ -2,10 +2,18 @@ Codion Change Log
 =================
 
 ## 0.18.30
+### is.codion.common.model
+- FilterModel.SortModel added along with sort().
+- FilterModel.VisibleItems.sort() renamed sortItems().
 ### is.codion.swing.common.ui
 - CompletionDocument bug fixed, prevented combo boxes from selecting a null value, if the model did not contain null.
 - ComboBoxMouseWheelListener bug fixed, now selects first item instead of second on down scroll with null selected, if the model did not contain null.
 - CompletionDocument.remove() bug fixed, did not handle empty combo box.
+### is.codion.framework.model
+- EntityTableModel.orderQueryBySortOrder() added.
+- AbstractEntityTableModel.orderByFromSortModel() added.
+### is.codion.swing.framework.ui
+- EntityTablePanel.orderQueryBySortOrder() moved to model.
 
 ## 0.18.29
 ### is.codion.common.model